requirements.txt
gradio
gtts
elevenlabs
python-dotenv
pydub
SpeechRecognition
groq